Fixed conflict and fixed bug in x3d_vex3d_fp0x16_dot

+++ src/headers/X3D_fix.h
@@ -56,7 +56,7 @@ typedef uint32_t uint32;
 typedef int64_t int64;
 
 /// A 64-bit unsigned integer
-typedef uint16_t uint64;
+typedef uint64_t uint64;
 
 #endif
